RAINGAUGE/requirements.txt
RPi.GPIO==0.7.0
APScheduler==3.7.0
six==1.15.0
tzlocal==2.1
pytz==2021.1
paho-mqtt~=1.5.1